SahtechAI/fastapi_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
groq==0.23.1
python-dotenv==1.0.1
httpx==0.26.0
